Consolidating Joystick functionality...
I now use a dummy joystick to poll for input from a newly connected
joystick. Important changes:

- The dummy joystick takes advantage of newly added Joystick
functionality that allows users to get input without mapping to a
JoystickConfig (e.g. I can get the first axis of a joystick now,
regardless its config.) This was not easily possible before.

- The JoystickManager now runs on Start instead of Awake. This way,
JoystickListeners can have priority when it comes to subscribing to its
events before they're fired. Not a perfect solution, but it works for
now.

- Renamed the "IsActive" property of Joysticks to IsConnected to comply
with prior renaming of the JoystickManager's events.